+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to the "git-hooks" extension will be documented in this file
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+### [1.3.0] -- 2023-07-29
+- Introduced `GitHooks` output channel to display the output logs of the extension
+- avoid await inside `activate` function of `extension.ts` to improve startup time
 
 ### [1.2.1] -- 2023-07-28
 - Bug fix for `git_hooks_scm` icon display in `ActivityBar`(SCM view badge was removed because of this bug)